requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyJWT>=2.8.0
numpy
opencv-python-headless
requests
deep-translator
Pillow
torch
